[Update] 6th compiled program

+++ GuiApp/CVPluginManager.cpp
@@ -31,7 +31,7 @@ const std::string &CVPluginManager::GetPluginAbsPath(CVPlugin *plugin)
 
 }
 
-size_t CVPluginManager::GetNumPlugins() const
+int CVPluginManager::GetNumPlugins() const
 {
     return mPlugins.size();
 }
